requirements.txt
aiofiles==23.1.0
aiogram==3.0.0rc2
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
async-timeout==4.0.3
attrs==23.1.0
Babel==2.9.1
certifi==2023.7.22
charset-normalizer==3.2.0
coloredlogs==15.0.1
executor==23.2
fasteners==0.18
frozenlist==1.4.0
humanfriendly==10.0
idna==3.4
magic-filter==1.0.11
multidict==6.0.4
property-manager==3.0
pydantic==2.2.1
pydantic_core==2.6.1
python-dotenv==1.0.0
pytz==2023.3
requests==2.31.0
six==1.16.0
typing_extensions==4.7.1
urllib3==2.0.4
verboselogs==1.7
yarl==1.9.2
